added uint256 in contract

+++ contracts/SimpleStorage.sol
@@ -3,9 +3,9 @@ pragma solidity >=0.4.21 <0.7.0;
 contract SimpleStorage {
     event StorageSet(string _message);
 
-    uint public storedData;
+    uint256 public storedData;
 
-    function set(uint x) public {
+    function set(uint256 x) public {
         storedData = x;
 
         emit StorageSet("Data stored successfully!");
